Navigation
Ive had some problems with the navigation. First the Button.js used for the media query and establishing the buttons in navigation would not allow page rendering and onclicks. I tried changing them in order to handle tab navigation but the buttons would not handle click events. I also tried to change them by setting them in <a></a> but that did not work either. I ended up creating a header and separating the navigation then using the App.js to create render page objects to handle tab changes.